Update index.html
Fixed the combinations for several aspects as they were incorrect. Also having the translated versions in the combinations makes it very confusing when trying to check if they are right and then to make corrections. Also I don't have the time atm to put them in alphabetical order but that would make checking them easier also.